+++ src/main/java/gpse/team52/service/LocationServiceImpl.java
@@ -20,7 +20,7 @@ public class LocationServiceImpl implements LocationService {
     }
 
     @Override
-    public Location createLocation(String name, int time) {
+    public Location createLocation(String name, long time) {
         final Location location = new Location(name);
         return createLocation(location);
     }
